# PERF: imporves performance in SeriesGroupBy.value_counts