```-------------------------------------------------------------------------------
Test name                                    | head[ms] | base[ms] |  ratio   |
-------------------------------------------------------------------------------
groupby_ngroups_10000_value_counts           |  11.4880 | 6698.5627 |   0.0017 |
groupby_ngroups_100_value_counts             |   1.3397 |  69.8457 |   0.0192 |
-------------------------------------------------------------------------------
Test name                                    | head[ms] | base[ms] |  ratio   |
-------------------------------------------------------------------------------

Ratio < 1.0 means the target commit is faster then the baseline.
Seed used: 1234

Target [d587be8] : PERF: imporves performance in SeriesGroupBy.value_counts
Base   [a3c4b59] : TST: pythonxs link seems to have changed in test_html.py, skip tests
```